+++ stdlib/public/runtime/ProtocolConformance.cpp
@@ -16,10 +16,10 @@
 
 #include "swift/Basic/LLVM.h"
 #include "swift/Basic/Lazy.h"
-#include "swift/Basic/Unreachable.h"
 #include "swift/Runtime/Concurrent.h"
 #include "swift/Runtime/Metadata.h"
 #include "swift/Runtime/Mutex.h"
+#include "swift/Runtime/Unreachable.h"
 #include "ImageInspection.h"
 #include "Private.h"
